requirements.txt
flask
pymongo
pillow
python-dotenv